+++ test/yulPhaser/TestHelpers.h
@@ -158,9 +158,9 @@ double mean(std::vector<T> const& _samples)
 
 	double sum = 0;
 	for (T const& sample: _samples)
-		sum += static_cast<double>(sample);
+		sum += double(sample);
 
-	return sum / _samples.size();
+	return sum / double(_samples.size());
 }
 
 /// Calculates the sum of squared differences between @a _expectedValue and the values of a series
@@ -179,9 +179,9 @@ double meanSquaredError(std::vector<T> const& _samples, double _expectedValue)
 
 	double sumOfSquaredDifferences = 0;
 	for (T const& sample: _samples)
-		sumOfSquaredDifferences += (sample - _expectedValue) * (sample - _expectedValue);
+		sumOfSquaredDifferences += (double(sample) - _expectedValue) * (double(sample) - _expectedValue);
 
-	return sumOfSquaredDifferences / _samples.size();
+	return sumOfSquaredDifferences / double(_samples.size());
 }
 
 }
